+++ bus-mapping/src/circuit_input_builder/execution.rs
@@ -98,9 +98,39 @@ impl ExecStep {
     }
 
     /// get rw index
-    pub fn rw_index(&self, index: usize) -> (Target, usize) {
-        let x = self.bus_mapping_instance[index];
-        (x.0, x.1)
+    pub fn rw_index(&self, index: usize) -> OperationRef {
+        self.bus_mapping_instance[index]
+    }
+
+    /// Get the size of read and writes
+    pub fn rw_indices_len(&self) -> usize {
+        self.bus_mapping_instance.len()
+    }
+
+    /// Get the spent gas
+    pub fn gas_spent(&self) -> u64 {
+        self.gas_left.0 - self.gas_cost.0
+    }
+
+    /// Get stack pointer
+    pub fn stack_pointer(&self) -> u64 {
+        1024 - self.stack_size as u64
+    }
+
+    /// The memory size in word **before** this step
+    pub fn memory_word_size(&self) -> u64 {
+        let n_bytes_word = 32u64;
+        let memory_size = self.memory_size as u64;
+        // EVM always pads the memory size to word size
+        // https://github.com/ethereum/go-ethereum/blob/master/core/vm/interpreter.go#L212-L216
+        // Thus, the memory size must be a multiple of 32 bytes.
+        assert_eq!(memory_size % n_bytes_word, 0);
+        memory_size / n_bytes_word
+    }
+
+    /// Get program counter
+    pub fn program_counter(&self) -> u64 {
+        self.pc.0.try_into().expect("program counter can fit u64")
     }
 }
