+++ eth-types/src/evm_types.rs
@@ -60,6 +60,12 @@ impl fmt::Debug for Gas {
     }
 }
 
+impl From<Gas> for u64 {
+    fn from(value: Gas) -> Self {
+        value.0
+    }
+}
+
 /// Quotient for max refund of gas used
 pub const MAX_REFUND_QUOTIENT_OF_GAS_USED: usize = 5;
 /// Gas stipend when CALL or CALLCODE is attached with value.
