+++ zkevm-circuits/src/evm_circuit/execution.rs
@@ -4,6 +4,7 @@ use super::{
         FIXED_TABLE_LOOKUPS, KECCAK_TABLE_LOOKUPS, N_BYTE_LOOKUPS, N_COPY_COLUMNS,
         N_PHASE1_COLUMNS, RW_TABLE_LOOKUPS, TX_TABLE_LOOKUPS,
     },
+    step::HasExecutionState,
     util::{instrumentation::Instrument, CachedRegion, CellManager, StoredExpression},
 };
 use crate::{
@@ -927,7 +928,7 @@ impl<F: Field> ExecutionConfig<F> {
                     if next.is_none() {
                         break;
                     }
-                    let height = step.execution_state.get_step_height();
+                    let height = step.execution_state().get_step_height();
 
                     // Assign the step witness
                     self.assign_exec_step(
@@ -1074,8 +1075,8 @@ impl<F: Field> ExecutionConfig<F> {
             return Ok(());
         }
         assert_eq!(height, 1);
-        assert!(step.rw_indices.is_empty());
-        assert!(matches!(step.execution_state, ExecutionState::EndBlock));
+        assert!(step.rw_indices_len() == 0);
+        assert!(matches!(step.execution_state(), ExecutionState::EndBlock));
 
         // Disable access to next step deliberately for "repeatable" step
         let region = &mut CachedRegion::<'_, '_, F>::new(
@@ -1088,7 +1089,7 @@ impl<F: Field> ExecutionConfig<F> {
         self.assign_exec_step_int(region, offset_begin, block, transaction, call, step)?;
 
         region.replicate_assignment_for_range(
-            || format!("repeat {:?} rows", step.execution_state),
+            || format!("repeat {:?} rows", step.execution_state()),
             offset_begin + 1,
             offset_end,
         )?;
@@ -1109,11 +1110,11 @@ impl<F: Field> ExecutionConfig<F> {
         next: Option<(&Transaction, &Call, &ExecStep)>,
         challenges: &Challenges<Value<F>>,
     ) -> Result<(), Error> {
-        if !matches!(step.execution_state, ExecutionState::EndBlock) {
+        if !matches!(step.execution_state(), ExecutionState::EndBlock) {
             log::trace!(
                 "assign_exec_step offset: {} state {:?} step: {:?} call: {:?}",
                 offset,
-                step.execution_state,
+                step.execution_state(),
                 step,
                 call
             );
@@ -1165,7 +1166,9 @@ impl<F: Field> ExecutionConfig<F> {
             };
         }
 
-        match step.execution_state {
+        let execution_state = ExecutionState::from(step);
+
+        match execution_state {
             // internal states
             ExecutionState::BeginTx => assign_exec_step!(self.begin_tx_gadget),
             ExecutionState::EndTx => assign_exec_step!(self.end_tx_gadget),
@@ -1304,7 +1307,7 @@ impl<F: Field> ExecutionConfig<F> {
                 assign_exec_step!(self.error_return_data_out_of_bound)
             }
 
-            _ => evm_unimplemented!("unimplemented ExecutionState: {:?}", step.execution_state),
+            _ => evm_unimplemented!("unimplemented ExecutionState: {:?}", execution_state),
         }
 
         // Fill in the witness values for stored expressions
@@ -1312,8 +1315,8 @@ impl<F: Field> ExecutionConfig<F> {
 
         // enable with `RUST_LOG=debug`
         if log::log_enabled!(log::Level::Debug) {
-            let is_padding_step = matches!(step.execution_state, ExecutionState::EndBlock)
-                && step.rw_indices.is_empty();
+            let is_padding_step =
+                matches!(execution_state, ExecutionState::EndBlock) && step.rw_indices_len() == 0;
             if !is_padding_step {
                 // expensive function call
                 Self::check_rw_lookup(
@@ -1337,8 +1340,8 @@ impl<F: Field> ExecutionConfig<F> {
         let mut assigned_stored_expressions = Vec::new();
         for stored_expression in self
             .stored_expressions_map
-            .get(&step.execution_state)
-            .unwrap_or_else(|| panic!("Execution state unknown: {:?}", step.execution_state))
+            .get(&step.execution_state())
+            .unwrap_or_else(|| panic!("Execution state unknown: {:?}", step.execution_state()))
         {
             let assigned = stored_expression.assign(region, offset)?;
             assigned.map(|v| {
@@ -1374,7 +1377,7 @@ impl<F: Field> ExecutionConfig<F> {
         }
 
         let rlc_assignments: BTreeSet<_> = step
-            .rw_indices
+            .bus_mapping_instance
             .iter()
             .map(|rw_idx| block.rws[*rw_idx])
             .map(|rw| {
@@ -1396,10 +1399,10 @@ impl<F: Field> ExecutionConfig<F> {
         // Check that the number of rw operations generated from the bus-mapping
         // correspond to the number of assigned rw lookups by the EVM Circuit
         // plus the number of rw lookups done by the copy circuit.
-        if step.rw_indices.len() != assigned_rw_values.len() + step.copy_rw_counter_delta as usize {
+        if step.rw_indices_len() != assigned_rw_values.len() + step.copy_rw_counter_delta as usize {
             log::error!(
                 "step.rw_indices.len: {} != assigned_rw_values.len: {} + step.copy_rw_counter_delta: {} in step: {:?}", 
-                step.rw_indices.len(),
+                step.rw_indices_len(),
                 assigned_rw_values.len(),
                 step.copy_rw_counter_delta,
                 step
@@ -1423,23 +1426,21 @@ impl<F: Field> ExecutionConfig<F> {
             // corresponding rw lookup, but in the bus-mapping they are
             // generated at the end of the step.
             let idx = if is_rev {
-                step.rw_indices.len() - rev_count
+                step.rw_indices_len() - rev_count
             } else {
                 idx - rev_count
             };
-            let rw_idx = step.rw_indices[idx];
-            let rw = block.rws[rw_idx];
+            let rw = block.get_rws(step, idx);
             let table_assignments = rw.table_assignment_aux(evm_randomness);
             let rlc = table_assignments.rlc(lookup_randomness);
             if rlc != assigned_rw_value.1 {
                 log::error!(
-                    "incorrect rw witness. lookup input name: \"{}\"\nassigned={:?}\nrlc     ={:?}\nrw index: {:?}, {}th rw of step {:?}, rw: {:?}",
+                    "incorrect rw witness. lookup input name: \"{}\"\nassigned={:?}\nrlc     ={:?}\n{}th rw of step {:?}, rw: {:?}",
                     assigned_rw_value.0,
                     assigned_rw_value.1,
                     rlc,
-                    rw_idx,
                     idx,
-                    step.execution_state,
+                    step.execution_state(),
                     rw);
             }
         }
