+++ zkevm-circuits/src/evm_circuit/execution/sstore.rs
@@ -184,24 +184,23 @@ impl<F: Field> ExecutionGadget<F> for SstoreGadget<F> {
             ),
         )?;
 
-        let [key, value] =
-            [step.rw_indices[5], step.rw_indices[6]].map(|idx| block.rws[idx].stack_value());
+        let [key, value] = [5, 6].map(|idx| block.get_rws(step, idx).stack_value());
         self.phase2_key
             .assign(region, offset, region.word_rlc(key))?;
         self.phase2_value
             .assign(region, offset, region.word_rlc(value))?;
 
-        let (_, value_prev, _, original_value) = block.rws[step.rw_indices[7]].storage_value_aux();
+        let (_, value_prev, _, original_value) = block.get_rws(step, 7).storage_value_aux();
         self.phase2_value_prev
             .assign(region, offset, region.word_rlc(value_prev))?;
         self.phase2_original_value
             .assign(region, offset, region.word_rlc(original_value))?;
 
-        let (_, is_warm) = block.rws[step.rw_indices[8]].tx_access_list_value_pair();
+        let (_, is_warm) = block.get_rws(step, 8).tx_access_list_value_pair();
         self.is_warm
             .assign(region, offset, Value::known(F::from(is_warm as u64)))?;
 
-        let (tx_refund, tx_refund_prev) = block.rws[step.rw_indices[9]].tx_refund_value_pair();
+        let (tx_refund, tx_refund_prev) = block.get_rws(step, 9).tx_refund_value_pair();
         self.tx_refund_prev
             .assign(region, offset, Value::known(F::from(tx_refund_prev)))?;
 
@@ -209,7 +208,7 @@ impl<F: Field> ExecutionGadget<F> for SstoreGadget<F> {
             region,
             offset,
             Value::known(F::from(GasCost::SSTORE_SENTRY.0)),
-            Value::known(F::from(step.gas_left)),
+            Value::known(F::from(step.gas_left.into())),
         )?;
 
         self.gas_cost
