+++ zkevm-circuits/src/evm_circuit/execution/swap.rs
@@ -70,7 +70,7 @@ impl<F: Field> ExecutionGadget<F> for SwapGadget<F> {
         self.same_context.assign_exec_step(region, offset, step)?;
 
         for (cell, value) in self.phase2_values.iter().zip(
-            [step.rw_indices[0], step.rw_indices[1]]
+            [step.rw_index(0), step.rw_index(1)]
                 .map(|idx| block.rws[idx].stack_value())
                 .iter(),
         ) {
